requirements.txt
pandas
python-dotenv
transformers
requests
openpyxl
torch
numpy<2.0